+++ CHANGELOG.md
@@ -9,10 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.0] - 2023-11-29
 ### Added
-- `@rarimo/rarime` - VC storage tests
+- `@rarimo/rarime` - credentials storage tests
 
 ### Changed
-- `@rarimo/rarime` - VC helpers
+- `@rarimo/rarime`
+  - refactored credential-helpers methods to `VCManager` class
+    - ceramic account injection
+    - unified ceramic client
+    - hash and encrypt data
+  - Moved credentials to ComposeDB
+
+## [1.0.2] - 2023-12-08
+### Changed
+- `@rarimo/rarime` - added new dashboard domains to the whitelist;
 
 ## [1.0.1] - 2023-11-29
 ### Fixed
@@ -128,7 +137,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented `@rarimo/rarime-connector` and `@rarimo/rarime` packages
 
 [Unreleased]: https://github.com/rarimo/rarime/compare/1.1.0...HEAD
-[1.1.0]: https://github.com/rarimo/rarime/compare/1.0.1...1.1.0
+[1.1.0]: https://github.com/rarimo/rarime/compare/1.0.2...1.1.0
+[1.0.2]: https://github.com/rarimo/rarime/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/rarimo/rarime/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/rarimo/rarime/compare/0.8.0...1.0.0
 [0.8.0]: https://github.com/rarimo/rarime/compare/0.7.2...0.8.0